# lab11/GO11_01/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST-интерфейс для работы с коллекцией Celebrities.",
        "title": "GO11_01 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/Celebrities": {
            "post": {
                "description": "Добавить элемент в коллекцию Celebrities. При дублировании id возвращается 409.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Celebrities"
                ],
                "summary": "Добавить элемент",
                "parameters": [
                    {
                        "description": "Данные элемента",
                        "name": "celebrity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Celebrity"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Элемент создан",
                        "schema": {
                            "$ref": "#/definitions/main.Celebrity"
                        }
                    },
                    "400": {
                        "description": "Некорректный JSON",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Элемент с таким id уже существует",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/Celebrities/All": {
            "get": {
                "description": "Получить все элементы коллекции Celebrities.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Celebrities"
                ],
                "summary": "Вся коллекция",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Celebrity"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/Celebrities/{id}": {
            "get": {
                "description": "Получить элемент коллекции Celebrities по id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Celebrities"
                ],
                "summary": "Элемент коллекции",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id элемента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/main.Celebrity"
                        }
                    },
                    "404": {
                        "description": "Элемент не найден",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Изменить элемент коллекции Celebrities по id. При отсутствии элемента возвращается 404.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Celebrities"
                ],
                "summary": "Изменить элемент",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id элемента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные элемента",
                        "name": "celebrity",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Celebrity"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/main.Celebrity"
                        }
                    },
                    "400": {
                        "description": "Некорректный JSON",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Элемент не найден",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удалить элемент коллекции Celebrities по id. При отсутствии элемента возвращается 404.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Celebrities"
                ],
                "summary": "Удалить элемент",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id элемента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/main.Celebrity"
                        }
                    },
                    "404": {
                        "description": "Элемент не найден",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Celebrity": {
            "type": "object",
            "properties": {
                "fullName": {
                    "type": "string",
                    "example": "Charlie Chaplin"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "nationality": {
                    "type": "string",
                    "example": "United Kingdom"
                },
                "reqPhotoPath": {
                    "type": "string",
                    "example": "photos/charlie_chaplin.jpg"
                }
            }
        }
    }
}